refactor: combine interfaces
