--- src/interface/interface.ts
@@ -1,3 +1,0 @@
-export interface ITimeTable {
-  [key: string]: boolean[];
-}
